requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
